convert c_double to c_int

+++ analysis/variables/src/FlavorTaggingVariables.cc
@@ -2271,7 +2271,7 @@ If one particle in the RestOfEvent is found to belong to the reconstructed ``B0`
 )DOC");
     REGISTER_METAVARIABLE("isRelatedRestOfEventMajorityB0Flavor(maskName)", isRelatedRestOfEventMajorityB0FlavorWithMask,
                       "[Expert] Returns 0 (1) if the majority of tracks and clusters of the RestOfEvent related to the given Particle are related to a ``anti-B0`` (``B0``).",
-                      Manager::VariableDataType::c_double);
+                      Manager::VariableDataType::c_int);
     REGISTER_VARIABLE("isRestOfEventMajorityB0Flavor", isRestOfEventMajorityB0Flavor,
                       "[Expert] Returns 0 (1) if the majority of tracks and clusters of the current RestOfEvent are related to a ``anti-B0`` (``B0``).");
     REGISTER_VARIABLE("mcFlavorOfOtherB", mcFlavorOfOtherB,  R"DOC(
